% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{async}
\alias{async}
\alias{async-package}
\title{Asynchronous HTTP}
\description{
Define tasks, that can stop on an HTTP request, and then continue later,
once the HTTP response is in. An event loop keeps the tasks running.
}
