% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async-apply.R
\name{async_apply}
\alias{async_apply}
\title{Creates an asynchronous function with some arguments already applied.}
\usage{
async_apply(func, ...)
}
\arguments{
\item{func}{The function you want to eventually apply all arguments
to.}

\item{...}{The arguments to apply to \code{func}.}
}
\value{
The partially applied function.
}
\description{
Useful as a shorthand when combined with other control flow
functions. Any arguments passed to the returned function are added to
the arguments originally passed to \code{async_apply}.
}
\examples{
f <- asyncify(function(a, b, c) a + b + c)
error <- result <- NULL
wait_for(parallel(
  list(
    async_apply(f, 1, 2, 3),
    async_apply(f, 4, 5, 6),
    async_apply(f, 7, 8, 9)
  ),
  function(err, res) { error <<- err; result <<- res }
))
result
}
\seealso{
Other async utilities: \code{\link{asyncify}},
  \code{\link{reflect}}, \code{\link{syncify}}
}
