% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async_bar.R
\name{async_bar}
\alias{async_bar}
\title{R6 Class async_bar}
\description{
Progress shiny bar to async class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{async_bar$new()}}
\item \href{#method-progress}{\code{async_bar$progress()}}
\item \href{#method-close}{\code{async_bar$close()}}
\item \href{#method-clone}{\code{async_bar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
: Create new progress bar.


: Progress Bar Inc.


: Set message to progress bar.


: Show progres bar.


: Hide progress bar.


: Remove progress bar.


: Interrupt progress update in client side.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async_bar$new(async, id, interval = 400, detail = "", session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{: ID of progress bar.}

\item{\code{id}}{: ID of progress bar.}

\item{\code{id}}{: ID of progress bar.}

\item{\code{id}}{: ID of progress bar.}

\item{\code{id}}{: ID of progress bar.}

\item{\code{id}}{: ID of progress bar.}

\item{\code{detail}}{: detail message to display.}

\item{\code{detail}}{: detail message to display.}

\item{\code{session}}{: shiny session.}

\item{\code{message}}{: Main message to display.}

\item{\code{show}}{: TRUE/FALSE.}

\item{\code{width}}{: width value.}

\item{\code{message}}{: Main message to display.}

\item{\code{width}}{: value for width bar.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
: js command.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-progress"></a>}}
\if{latex}{\out{\hypertarget{method-progress}{}}}
\subsection{Method \code{progress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async_bar$progress(session, input)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async_bar$close(session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async_bar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
