% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async-call.R
\name{async_call}
\alias{async_call}
\title{Make an async function call}
\usage{
async_call(fun, args, callback)
}
\arguments{
\item{fun}{Function to call, a function object.}

\item{args}{Other arguments to pass to the function.}

\item{callback}{Function to pass as the callback (last) argument of
\code{fun}.}
}
\value{
The return value of the called \code{fun}.
}
\description{
We need to do some extra work, compared to a regular function call,
because we need to work out the name of the callback argument, which is
the last argument of the function.
}
\details{
This is because the function might take variable number of arguments via
\code{...} and then we have to name the callback argument. Or, some argument
(probably with default values) are not included in the call, but we
still want to make sure that the callback is passed to the last
argument.
}
\keyword{internal}
