% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deferred.R
\name{async_cancel_pending}
\alias{async_cancel_pending}
\title{Cancel async tasks}
\usage{
async_cancel_pending(..., .list = list())
}
\arguments{
\item{...}{Deferred values to cancel.}

\item{.list}{More deferred values to cancel.}
}
\description{
Does nothing for deferred values that are already resolved.
Does nothing for objects that are not deferred values.
}
