% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\name{async_constant}
\alias{async_constant}
\title{Make a minimal deferred that resolves to the specified value}
\usage{
async_constant(value = NULL)
}
\arguments{
\item{value}{The value to resolve to.}
}
\value{
A deferred value.
}
\description{
This is sometimes useful to start a deferred chain.
}
\examples{
dx <- async_constant(1/100)$
  then(~ delay(.))$
  then(~ print(.))
await(dx)
}
