% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyncBar1.R
\name{asyncBar1}
\alias{asyncBar1}
\title{R6 Class asyncBar1}
\description{
Reports  async class progress with a default shiny bar to end user.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-asyncBar1-new}{\code{asyncBar1$new()}}
\item \href{#method-asyncBar1-progress}{\code{asyncBar1$progress()}}
\item \href{#method-asyncBar1-finalize}{\code{asyncBar1$finalize()}}
\item \href{#method-asyncBar1-clone}{\code{asyncBar1$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-asyncBar1-new"></a>}}
\if{latex}{\out{\hypertarget{method-asyncBar1-new}{}}}
\subsection{Method \code{new()}}{
Interactive environment
to create a progress  bar of async tracking.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{asyncBar1$new(async, id, interval = 1000, max.rep = 50)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{async}}{R6 class. Object to tracking routines.}

\item{\code{id}}{character. ID of progress bar.}

\item{\code{interval}}{numeric. Approximate number of milliseconds
to wait between checks of the job progress.}

\item{\code{max.rep}}{numeric. Maximum number of times that
a progress value can be repeated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-asyncBar1-progress"></a>}}
\if{latex}{\out{\hypertarget{method-asyncBar1-progress}{}}}
\subsection{Method \code{progress()}}{
observe event to checks of the async job progress.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{asyncBar1$progress(session, input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{shiny session}

\item{\code{input}}{shiny input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-asyncBar1-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-asyncBar1-finalize}{}}}
\subsection{Method \code{finalize()}}{
Close all routines of async bar.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{asyncBar1$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-asyncBar1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-asyncBar1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{asyncBar1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
