% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyncBar2.R
\name{asyncBar2}
\alias{asyncBar2}
\title{R6 Class asyncBar1}
\description{
Reports  async class progress with a custom shiny bar to end user.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{asyncBar2$new()}}
\item \href{#method-progress}{\code{asyncBar2$progress()}}
\item \href{#method-cancel}{\code{asyncBar2$cancel()}}
\item \href{#method-finalize}{\code{asyncBar2$finalize()}}
\item \href{#method-clone}{\code{asyncBar2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Interactive environment
to create a progress  bar of async tracking.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{asyncBar2$new(async, id, interval = 1000, max.rep = 50, msg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{async}}{R6 class. Object to tracking routines.}

\item{\code{id}}{character. ID of progress bar.}

\item{\code{interval}}{numeric. Approximate number of milliseconds
to wait between checks of the job progress.}

\item{\code{max.rep}}{numeric. Maximum number of times that
a progress value can be repeated.}

\item{\code{msg}}{A single-element character vector;
the message to be displayed to the user,}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-progress"></a>}}
\if{latex}{\out{\hypertarget{method-progress}{}}}
\subsection{Method \code{progress()}}{
observe event to checks of the async job progress.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{asyncBar2$progress(session, input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{shiny session}

\item{\code{input}}{shiny input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-cancel}{}}}
\subsection{Method \code{cancel()}}{
display  a cancel button to end user
associated with the process tracked by the async class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{asyncBar2$cancel(session, input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{shiny session}

\item{\code{input}}{shiny input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Close all routines of async bar.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{asyncBar2$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{asyncBar2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
