% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async-description.R
\name{asyncOpts}
\alias{asyncOpts}
\title{Get or set global options for the async package. These options are
mostly used in package development or unit testing.}
\usage{
asyncOpts(
  verbose = get("verbose", parent.env(environment())),
  compileLevel = get("compileLevel", parent.env(environment())),
  paranoid = get("paranoid", parent.env(environment())),
  destructive = get("destructive", parent.env(environment()))
)
}
\arguments{
\item{verbose}{if TRUE, coroutines will print an extremely chatty
play-by-play of what they are doing.}

\item{compileLevel}{Default compile level for new generators. See
\url{gen}.}

\item{paranoid}{if TRUE, perform various time-consuming integrity
checks on the result of compilation. Meant to be enabled during
testing.}

\item{destructive}{defaults TRUE: tear down interpreted coroutines
while building compiled replacements. Used in package debugging.}
}
\value{
a list containing the current settings.
}
\description{
If new settings are provided in arguments they take effect for any
new generators or asyncs created.
}
