% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/await.R
\name{await}
\alias{await}
\title{Wait for some (or all) tasks to finish}
\usage{
await(ids = NULL)
}
\arguments{
\item{ids}{A vector of task ids, or \code{NULL}.}
}
\description{
If \code{ids} is \code{NULL} then it waits for all tasks to finish, i.e. the
event loop must be completely empty. This is rarely useful.
}
\details{
If \code{ids} is a a charater vector, then it waits for all listed tasks.
}
\examples{
id <- parallel(
  list(
    function(callback) { Sys.sleep(1/100); callback(NULL, 1) },
    function(callback) { Sys.sleep(1/200); callback(NULL, 2) },
    function(callback) {
      http_get("http://httpbin.org", function(err, res) {
        callback(err, res$status_code)
      })
    }
  ),
  function(err, res) print(res)
)
await(id)
}
