% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/await.R
\name{await_all}
\alias{await_all}
\title{Wait for a list of deferred values}
\usage{
await_all(..., .list = list())
}
\arguments{
\item{...}{Deferred values to wait on.}

\item{.list}{More deferred values to wait on.}
}
\description{
Similar to \code{\link[=await]{await()}}, but waits for a list of deferred values.
}
\examples{
afun <- async(function() {
  urls <- c("https://eu.httpbin.org?q=1", "https://eu.httpbin.org?q=2")
  dx <- lapply(urls, http_head)
  resp <- await_all(.list = dx)
  lapply(resp, "[[", "status_code")
})
synchronise(afun())
}
\seealso{
Other await functions: \code{\link{await_any}},
  \code{\link{await_env}}, \code{\link{await}}
}
\concept{await functions}
