% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/await.R
\name{await_any}
\alias{await_any}
\title{Wait for any of the listed tasks to finish}
\usage{
await_any(ids)
}
\arguments{
\item{ids}{A (non-empty) list of tasks ids to wait on.}
}
\description{
Wait for any of the listed tasks to finish
}
\examples{
d1 <- http_get("http://httpbin.org/delay/1",
               function(err, res) print(res$url))
d2 <- http_get("http://httpbin.org/get",
               function(err, res) print(res$url))
await_any(c(d1, d2))
await(c(d1, d2))
}
\seealso{
Other synchronization functions: \code{\link{await}}
}
