% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/await_env.R
\name{await_env}
\alias{await_env}
\title{Wait for all deferred values in an environment}
\usage{
await_env(env)
}
\arguments{
\item{env}{Environment to wait on. An error is thrown if not an
environment.}
}
\value{
A named list of resolved values. An unhandled rejection make
\code{await_env} throw an error. Note that since the environment is a hash,
the order of the values is arbitrary.
}
\description{
The function returns when all deferred values are resolved (or
rejected).
}
\details{
It is permitted to change the environment (from within the async
functions, as the is no synchronous code running and R is single
threaded) while the wait is taking place. This allows waiting on a
dynamic set of async operations.
}
\examples{
## dynamically change the async operations we are waiting on
afun <- async(function() {
  env <- new.env()
  env$foo <- delay(1/10000)$
    then(function() {
      env$foo2 <- async_constant("OK2")
      "OK"
    })
  await_env(env)
})
sync_wrap(afun())
}
\seealso{
Other await functions: \code{\link{await_all}},
  \code{\link{await_any}}, \code{\link{await}}
}
\concept{await functions}
