% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call-function.R
\name{call_function}
\alias{call_function}
\title{Asynchronous function call, in a worker pool}
\usage{
call_function(func, args = list())
}
\arguments{
\item{func}{Function to call. See also the notes at \code{\link[callr:r]{callr::r()}}.}

\item{args}{Arguments to pass to the function. They will be copied
to the worker process.}
}
\value{
Deferred object.
}
\description{
The function will be called on another process, very much like
\code{\link[callr:r]{callr::r()}}.
}
