% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-loop.R
\name{call_with_callback}
\alias{call_with_callback}
\title{Call \code{func} and then call \code{callback} with the result}
\usage{
call_with_callback(func, callback)
}
\arguments{
\item{func}{Function to call.}

\item{callback}{Callback to call with the result of \code{func()},
or the error thrown.}
}
\description{
\code{callback} will be called with two arguments, the first one will the
error object if \code{func()} threw an error, or \code{NULL} otherwise. The second
argument is \code{NULL} on error, and the result of \code{func()} otherwise.
}
\keyword{internal}
