% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_functions.R
\name{ch_apply}
\alias{ch_apply}
\title{Apply a function to each element of a channel}
\usage{
ch_apply(ch, fun, ...)
}
\arguments{
\item{ch}{A \link{channel} object.}
}
\value{
A new \link{channel} object.
}
\description{
\code{ch_apply(ch, fun)} constructs a channel which applies \code{fun} to each element
received from \code{ch}.
}
\author{
Peter Meilstrup
}
