% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_functions.R
\name{ch_combine}
\alias{ch_combine}
\title{Combine several channels or promises into one channel.}
\usage{
ch_combine(...)
}
\arguments{
\item{...}{Each argument should be a \link{promise} or a \link{channel}.}
}
\value{
a \link{channel} object.
}
\description{
\code{ch_combine(...)} takes any number of \link{promise} or \link{channel}
objects. It awaits each one, and returns a \link{channel} object
which re-emits every value from its targets, in whatever
order they are received.
}
\author{
Peter Meilstrup
}
