% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect.channel}
\alias{collect.channel}
\title{Collect all values emitted by a channel.}
\usage{
collect.channel(ch, type = list())
}
\arguments{
\item{ch}{a \link{channel} object.}

\item{type}{Optionally provide a vector of the desired output type
(similarly to using \code{vapply}); defaults to \code{list()}}
}
\description{
\code{collect.channel} is analagous to \code{as.list.iteror}; it takes in a
channel and returns a promise. When the channel closes, the promise
will resolve with a vector containing all items returned by the channel.
}
