% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\title{Execute a function and collect a set of values by callback.}
\usage{
collect(fn, type = list())
}
\arguments{
\item{fn}{A function, which should accept a single argument "yield".}

\item{type}{A prototype output vector (i.e. a vector of the
same type as the desired output, similar to \link{vapply}.) Defaults
to \code{list()}.}
}
\value{
A vector of the same type.
}
\description{
\code{collect} calls the function in its argument, supplying a callback
\code{yield(val, name=NULL)}. Each value passed to \code{yield} is collected
and the list of all values is returned after \code{fn} returns.
}
\examples{

#cumulative sum
cumsum <- function(vec) {
  total <- 0
  collect(type=0, function(yield) {
    for (i in vec) total <- yield(total+i)
  }
}
}
\author{
Peter Meilstrup
}
