% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel.R
\name{combine}
\alias{combine}
\title{Combine several channels into one.}
\usage{
combine(...)
}
\arguments{
\item{...}{Each argument should be a \link{promise} or a \link{channel}.}
}
\value{
a \link{channel} object.
}
\description{
\code{combine(...)} takes any number of \link{promise} or \link{channel}
objects. It awaits each one, and returns a \link{channel} object
which re-emits every value from its targets, in whatever
order they are received.
}
\author{
Peter Meilstrup
}
