% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Compile a generator into a more efficient form.}
\usage{
compile(x, level, ...)
}
\arguments{
\item{x}{a \href{gen}{generator} object}

\item{level}{The compilation level. Currently implemented levels are:
0: Do nothing.}
}
\value{
A newly constructed generator. It may or may not share state variables with the input; using the input generator after compiling is not supported.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Peter Meilstrup
}
