% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coroutine.R
\name{debugAsync}
\alias{debugAsync}
\title{Toggle single-step debugging for a coroutine.}
\usage{
debugAsync(x, R, internal)
}
\arguments{
\item{x}{A coroutine object as constructed by (\link{async}, \link{gen} or \link{stream}).}

\item{R}{set TRUE to step through expressions at user level}

\item{internal}{Set TRUE to step through at coroutine implementation level.}
}
\value{
a `list(R=, internal=) with the current debug setting.
}
\description{
Toggle single-step debugging for a coroutine.
}
