% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deferred.R
\name{def__make_error_object}
\alias{def__make_error_object}
\title{Create an error object for a rejected deferred computation}
\usage{
def__make_error_object(self, private, err)
}
\arguments{
\item{self}{self}

\item{private}{private self}
}
\value{
error object
}
\description{
\itemize{
\item Make sure that the error is an error object.
\item Make sure that the error stack is in the correct format.
\item Make sure that the error has the correct classes.
}
}
\keyword{internal}
