% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay.R
\name{delay}
\alias{delay}
\title{Run a function after the specified time interval}
\usage{
delay(delay)
}
\arguments{
\item{delay}{Time interval in seconds, the amount of time to delay
to delay the execution of the callback. It can be a fraction of a
second.}
}
\value{
Task id, it can be waited on with \code{\link[=wait_for]{wait_for()}}.
}
\description{
Since R is single-threaded, the callback might be executed (much) later
than the specified time period.
}
\examples{
TODO
}
