% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detect}
\alias{detect}
\title{Returns the first value in a collection that passes an async truth test}
\usage{
detect(coll, iteratee, callback)
}
\arguments{
\item{coll}{A collection to iterate over.}

\item{iteratee}{A truth test to apply to each item in \code{coll}.
The \code{iteratee} must complete with a logical flag value as its result.
Invoked with \code{(item, callback)}.}

\item{callback}{A callback which is called as soon as any iteratee
returns true, or after all the \code{iteratee} functions have finished.
Result will be the first item in the array that passes the truth test
(iteratee) or \code{NULL} if none passed. Invoked with \code{(err, result)}.}
}
\value{
Task id.
}
\description{
The \code{iteratee} is applied in parallel, meaning the first \code{iteratee} to
return true will fire the \code{detect} callback with that result. That means
the result might not be the first item in the original \code{coll} (in terms
of order) that passes the test. If order within the original \code{coll} is
important, then look at \code{\link[=detect_series]{detect_series()}}.
}
\examples{
## Get an URL that works
id <- detect(
  c("https://eu.httpbin.org", "https://httpbin.org"),
  sequence(http_get, asyncify(function(x) x$status_code)),
  function(err, res) print(res)
)
await(id)
}
\seealso{
Other async iterators: \code{\link{amap}},
  \code{\link{detect_limit}}, \code{\link{detect_series}},
  \code{\link{each}}, \code{\link{every}},
  \code{\link{filter}}, \code{\link{some}}
}
