% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-series.R
\name{detect_series}
\alias{detect_series}
\title{Serial version of \code{detect}}
\usage{
detect_series(coll, iteratee, callback)
}
\arguments{
\item{coll}{A collection to iterate over.}

\item{iteratee}{A truth test to apply to each item in \code{coll}.
The \code{iteratee} must complete with a logical flag value as its result.
Invoked with \code{(item, callback)}.}

\item{callback}{A callback which is called as soon as any iteratee
returns true, or after all the \code{iteratee} functions have finished.
Result will be the first item in the array that passes the truth test
(iteratee) or \code{NULL} if none passed. Invoked with \code{(err, result)}.}
}
\value{
Task id.
}
\description{
The same as \code{\link[=detect]{detect()}} but runs only a single async operation at a time.
}
\examples{
nums <- NULL
result <- "not null"
await(detect_series(
  1:5,
  function(x, callback) { nums <<- c(nums, x) ; callback(NULL, FALSE) },
  function(err, res) result <<- res
))
nums
result
}
\seealso{
Other async iterators: \code{\link{amap}},
  \code{\link{detect_limit}}, \code{\link{detect}},
  \code{\link{each}}, \code{\link{every}},
  \code{\link{filter}}, \code{\link{some}}
}
