% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/each-of.R
\name{each_of}
\alias{each_of}
\title{Applies an iteratee to each item in collection, in parallel, with index}
\usage{
each_of(coll, iteratee, callback)
}
\arguments{
\item{coll}{A collection to iterate over.}

\item{iteratee}{An async function to apply to each item in \code{coll}.
Invoked with \code{(item, callback)}. The array index is not passed to the
iteratee. If  you need the index, use \code{\link[=each_of]{each_of()}}.}

\item{callback}{A callback which is called when all iteratee functions
have finished, or an error occurs. Invoked with \code{(err)}.}
}
\description{
Like each, except that it passes the index as the second argument to
the iteratee.
}
