% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\docType{data}
\name{event_emitter}
\alias{event_emitter}
\title{Generic Event Emitter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
event_emitter
}
\description{
TODO
}
\keyword{datasets}
