% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every.R
\name{every}
\alias{every}
\title{Returns \code{TRUE} if every element in collection satisfies an async test}
\usage{
every(coll, iteratee, callback)
}
\arguments{
\item{coll}{A collection to iterate over.}

\item{iteratee}{An async truth test to apply to each item in the
collection in parallel. The iteratee must complete with a logical
scalar result value. Invoked with \code{(item, callback)}.}

\item{callback}{A callback which is called after all the iteratee
functions have finished. Result will be either \code{TRUE} or \code{FALSE}
depending on the values of the async tests. Invoked with
\code{(err, result)}.}
}
\value{
Task id.
}
\description{
If any iteratee call returns \code{FALSE}, the main callback is immediately
called.
}
\examples{
## Check if all numbers are odd
result <- NULL
wait_for(every(
  c(1,3,5,7,10,11),
  function(x, callback) callback(NULL, as.logical(x \%\% 2)),
  function(err, res) result <<- res
))
result
}
\seealso{
Other async iterators: \code{\link{detect_limit}},
  \code{\link{detect_series}}, \code{\link{detect}},
  \code{\link{each}}, \code{\link{filter}},
  \code{\link{some}}
}
