% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onload.R
\name{get_default_event_loop}
\alias{get_default_event_loop}
\title{Default event loop of the R session.}
\usage{
get_default_event_loop()
}
\value{
The default event loop of the R session.
}
\description{
This event loop is created when the \code{async} package is loaded, and
all asyncronous constructs use this event loop by default.
}
\seealso{
\link{event_loop}
}
\keyword{internal}
