% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{http_head}
\alias{http_head}
\title{Asynchronous HTTP HEAD request}
\usage{
http_head(url)
}
\arguments{
\item{url}{URL to connect to.}
}
\value{
Task id that can be waited on with \code{\link[=wait_for]{wait_for()}}.
}
\description{
Asynchronous HTTP HEAD request
}
\examples{
TODO
}
\seealso{
Other asyncronous HTTP calls: \code{\link{http_get}}
}
