% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{http_setopt}
\alias{http_setopt}
\title{Set curl HTTP options in an event loop}
\usage{
http_setopt(total_con = NULL, host_con = NULL, multiplex = NULL)
}
\arguments{
\item{total_con, host_con, }{multiplex They are passed to
\code{\link[curl:multi]{curl::multi_set()}}. If an argument is \code{NULL} (the default) then it is
ignored.}
}
\description{
The event loop must be already running. In other words, you can only
call this function from async functions.
}
\details{
The default values are set when the first deferred HTTP operation of the
event loop is created, and they are taken from the \code{async_http_total_con},
\code{async_http_host_con} and \code{async_http_multiplex} options.
}
\seealso{
Other asyncronous HTTP calls: 
\code{\link{http_get}()},
\code{\link{http_head}()}
}
\concept{asyncronous HTTP calls}
