% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteror.R
\name{ilimit}
\alias{ilimit}
\title{Limit the number of elements emitted by an iterator.}
\usage{
ilimit(it, n)
}
\arguments{
\item{it}{an \link{iteror} os something that converts to an iteror.}

\item{n}{How many elements to take.}
}
\value{
a new \link{iteror} wrapping the provided one, which will stop
after the specified number of elements.
}
\description{
Limit the number of elements emitted by an iterator.
}
