% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deferred.R
\name{is_deferred}
\alias{is_deferred}
\title{Is object a deferred value?}
\usage{
is_deferred(x)
}
\arguments{
\item{x}{object}
}
\value{
Whether it is a deferred value.
}
\description{
Is object a deferred value?
}
\examples{
is_deferred(1:10)
afun <- function() {
  print(is_deferred(dx <- delay(1/100)))
  dx
}
synchronise(afun())
}
