% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteror.R
\name{iseq}
\alias{iseq}
\title{Iterator equivalent of \link{seq}.}
\usage{
iseq(from = 1L, to = Inf, by = 1L)
}
\arguments{
\item{from}{The starting value of the sequence.}

\item{to}{The value to stop at.}

\item{by}{How much to increment at each step.}
}
\value{
a new \link{iteror} which will yield the sequence specified.
}
\description{
Iterator equivalent of \link{seq}.
}
