% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_channel.R
\name{lazy_channel}
\alias{lazy_channel}
\title{an altername formulation to construct "channels".
This kind of channel is "lazy"
But it can also pass arguments to nextThen
(which is good for nextThen)}
\usage{
lazy_channel(handler, ..., max_awaiting = 500)
}
\arguments{
\item{`handler`}{should be a function which takes three arguments
\verb{\\(yield, reject, close)}. This function will ba called for
\emph{each} requested value; the contract is that once the handler is
called, you should later call either \code{emit(val)}, \code{reject(err)},
or \code{close()}}
}
\description{
an altername formulation to construct "channels".
This kind of channel is "lazy"
But it can also pass arguments to nextThen
(which is good for nextThen)
}
