% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_async}
\alias{load_async}
\title{save_async}
\usage{
load_async(file)
}
\arguments{
\item{file}{a connection or the name of the file where the R object is saved to or read from.}
}
\description{
Function to read a single async object saved as rds.
}
