% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{make_deferred_http}
\alias{make_deferred_http}
\title{Make a deferred object for an HTTP query}
\usage{
make_deferred_http(cb, file, on_progress)
}
\arguments{
\item{cb}{Callback function that should return a handle that can be
passed to \code{\link[curl:multi_add]{curl::multi_add()}}.}

\item{file}{Deprecated and should be \code{NULL}.}

\item{on_progress}{Deprecated and should be \code{NULL}.}
}
\value{
A deferred object.
}
\description{
Make a deferred object for an HTTP query
}
