% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue.R
\name{make_queue}
\alias{make_queue}
\title{Creates a queue object with the specified concurrency}
\usage{
make_queue(worker_function, concurrency = 1)
}
\arguments{
\item{worker_function}{Worker function. It will be called for each
item inserted in the queue. It will be called with two arguments,
the item, and a callback function. The callback function
is to be called once the worker finished its job. The callback
function has two arguments, the first one is used to report errors,
it \code{NULL} if no errors happened. The second argument is the result
of the worker's computation.}

\item{concurrency}{Number of parallel asynchronous workers allowed.}
}
\value{
A \link{queue} object, that can be used to manage the workers.
}
\description{
Items added to the queue are processed in parallel (up to the
concurrency limit). If all workers are in progress, the item is queued
until one becomes available. Once a worker completes work on an item,
that item's callback is called.
}
\seealso{
\link{queue} to manage the workers of a queue.
}
