% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{makeGraph}
\alias{makeGraph}
\title{Draw a graph representation of a generator.}
\usage{
makeGraph(x, file = "", sep = "\\n", ...)
}
\arguments{
\item{x}{A \link[=gen]{generator} or \link{async} object.}

\item{file}{Default will write to stdout. See \link{cat}.}

\item{sep}{Line separator; defaults to \code{"\\n"}. See \link{cat}.}

\item{...}{Other arguments to be passed to \link{cat}.}

\item{filename}{The file to write to.

makeGraph will write a Graphviz DOT format file describing the
given generator or async/await block. If you have Graphviz
installed you can then use the \code{dot} command to render the graph.}
}
\value{
a list containing information about the async/iterator.
}
\description{
Draw a graph representation of a generator.
}
\examples{
randomWalk <- gen({x <- 0; repeat {yield(x); x <- x + rnorm(1)}})
\dontrun{
#write a dotfile
makeGraph(randomWalk, "rwalk.dot")

#compile graph to ODF:
system("dot -Tpdf rwalk.dot > rwalk.pdf")

#or, display it in R with the Rgraphviz package:
g <- Rgraphviz::agread("rwalk.dot")
Rgraphviz::plot(g)
}
}
