% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parallel_limit}
\alias{parallel_limit}
\title{Run a list of functions in parallel, limiting the number of async tasks}
\usage{
parallel_limit(tasks, limit, callback)
}
\arguments{
\item{tasks}{A list (or collection) of async functions to run. Each
async function can complete with any number of optional result values.}

\item{limit}{The maximum number of async operations at a time.}

\item{callback}{A callback to run once all the functions have
completed successfully. This function gets a results list
containing all the result arguments passed to the task
callbacks. Invoked with \code{(err, results)}.}
}
\description{
The same as parallel but runs a maximum of limit async operations at a
time.
}
\examples{
error <- result <- NULL
wait_for(parallel_limit(
  list(
    function(callback) set_timeout(1/100, function() callback(NULL, 1)),
    function(callback) set_timeout(2/100, function() callback(NULL, 2)),
    function(callback) set_timeout(3/100, function() callback(NULL, 3))
  ),
  limit = 2,
  function(err, res) { error <<- err ; result <<- res }
))
error
result
}
\seealso{
Other async control flow: \code{\link{parallel}},
  \code{\link{retryable}}, \code{\link{retry}},
  \code{\link{sequence}}, \code{\link{until}},
  \code{\link{waterfall}}, \code{\link{whilst}}
}
