% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{reactiveTrigger}
\alias{reactiveTrigger}
\title{reactiveTrigger}
\usage{
reactiveTrigger()
}
\description{
A reactive trigger can be used when you want
to be able to explicitly trigger a reactive expression.
You can think of it as being similar to an action button,
except instead of clicking on a button to trigger an expression,
you can programatically cause the trigger.
This concept and code was created by Joe Cheng (author of shiny).
}
