% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{nextOr}
\alias{iteror}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{iterors}{\code{\link[iterors]{iteror}}, \code{\link[iterors]{nextOr}}}
}}

