% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retry.R
\name{retry}
\alias{retry}
\title{Retry a task a number of times}
\usage{
retry(task, times, callback)
}
\arguments{
\item{task}{An async function to retry. Invoked with \code{(callback)}.}

\item{times}{The number of attempts to make before giving up.}

\item{callback}{A callback which is called when the task has succeeded,
or after the final failed attempt. It receives the \code{err} and \code{result}
arguments of the last attempt at completing the task. Invoked with
\code{(err, results)}.}
}
\value{
Task id.
}
\description{
Attempts to get a successful response from task no more than \code{times}
times before returning an error. If the task is successful, the callback
will be passed the result of the successful task. If all attempts fail,
the callback will be passed the error and result (if any) of the final
attempt.
}
\examples{
## Try a download at most 5 times
result <- NULL
wait_for(retry(
  function(callback) http_get("https://httpbin.org", callback),
  times = 5,
  function(err, res) result <<- res$status_code
))
result
}
\seealso{
Other async control flow: \code{\link{parallel_limit}},
  \code{\link{parallel}}, \code{\link{retryable}},
  \code{\link{sequence}}, \code{\link{until}},
  \code{\link{waterfall}}, \code{\link{whilst}}
}
