% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_async}
\alias{save_async}
\title{save_async}
\usage{
save_async(async, file)
}
\arguments{
\item{async}{R6 class. Object to tracking routines.}

\item{file}{a connection or the name of the file where the R object is saved to or read from.
If its is missing a unique name file is created by default.}
}
\description{
Functions to write a single async object to a  RDS file.
}
