% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{sequence}
\alias{sequence}
\title{Creates a function which is a composition of asynchronous functions}
\usage{
sequence(..., .list = NULL)
}
\arguments{
\item{...}{Asynchronous functions to compose.}

\item{.list}{More asynchronous functions to compare. Appended to the
previous ones. This interface allows easier programming.}
}
\value{
Asynchronous function, the composition of all input functions.
}
\description{
Each function consumes the return value of the previous one.
Composing functions f(), g(), and h() would produce the result
of h(g(f())), only this version uses callbacks to obtain the return
values.
}
\examples{
http_status <- sequence(
  http_get,
  asyncify(function(x) x$status_code)
)
wait_for(http_status(
  "https://httpbin.org",
  callback = function(err, res) print(res)
))
}
\seealso{
Other async control flow: \code{\link{parallel_limit}},
  \code{\link{parallel}}, \code{\link{retryable}},
  \code{\link{retry}}, \code{\link{until}},
  \code{\link{waterfall}}, \code{\link{whilst}}
}
