% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeout.R
\name{set_timeout}
\alias{set_timeout}
\title{Run a function after the specified time interval}
\usage{
set_timeout(delay, callback)
}
\arguments{
\item{delay}{Time interval in seconds, the amount of time to delay
to delay the execution of the callback. It can be a fraction of a
second.}

\item{callback}{The function to call after \code{delay} seconds. It will be
called without arguments.}
}
\value{
Task id, it can be waited on with \code{\link[=wait_for]{wait_for()}}.
}
\description{
Since R is single-threaded, the callback might be executed (much) later
than the specified time period.
}
\examples{
result <- NULL
id <- set_timeout(1, function() result <<- "done")
result
wait_for(id)
result
}
