% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sockets.R
\name{channel.servsockconn}
\alias{channel.servsockconn}
\alias{channel.sockconn}
\title{Using socket connections through channels}
\usage{
\method{channel}{servsockconn}(obj, ...)

\method{channel}{sockconn}(
  obj,
  ...,
  read = {
     if (summary(obj)$text == "text") 
         c("lines")
     else c("bin",
    "lines", "char")
 },
  read_params = {
     switch(read, lines = list(n = 1), char = list(nchar = 1), bin =
    list(what = "raw", n = 1))
 }
)
}
\arguments{
\item{obj}{a \link[=socketConnection]{socket connection}.}

\item{read}{can be "lines", "obj", or "char"}

\item{read_params}{List of default parameters to pass to \code{readLines},
\code{readChar} or \code{readBin} as appropriate.}
}
\description{
\code{channel.servsockconn(obj)} takes an open server
socket \code{obj}, as returned from \link{serverSocket(...)}. and
constructs a channel which yields an opened \link{socketConnection}
object for each incoming connection.

\code{channel.sockconn(obj)} takes an open socket
connection and returns a channel which yields data as it comes
in.  \link{socketConnection} object for each incoming connection to
the socket.
}
