% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some.R
\name{some}
\alias{some}
\title{Returns \code{TRUE} if at least one element in a collection satisfies an
async test}
\usage{
some(coll, iteratee, callback)
}
\arguments{
\item{coll}{A collection to iterate over.}

\item{iteratee}{An async truth test to apply to each item in the
collections in parallel. The iteratee should complete with a boolean
result value. Invoked with \code{(item, callback)}.}

\item{callback}{A callback which is called as soon as any iteratee
returns \code{TRUE}, or after all the iteratee functions have finished.
Result will be either \code{TRUE} or \code{FALSE} depending on the values of the
async tests. Invoked with \code{(err, result)}.}
}
\value{
A task id, that can be waited on with \code{\link[=await]{await()}}.
}
\description{
If any iteratee call returns true, the main callback is
immediately called.
}
\examples{
## Check if one of these sites are up
await(some(
  c("https://eu.httpbin.org", "https://httpbin.org"),
  sequence(http_head, asyncify(function(x) x$status_code)),
  function(err, res) print(res)
))
}
\seealso{
Other async iterators: \code{\link{amap}},
  \code{\link{detect}}, \code{\link{each}},
  \code{\link{every}}, \code{\link{filter}}
}
