% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.r
\name{goto}
\alias{goto}
\title{Coroutine switch with delimited goto.}
\usage{
goto(branch = NULL)
}
\description{
The \code{switch} function implemented for coroutines in the \code{async}
package is more strict than the one in base R. Base R \code{switch} will
silently return NULL if no branch matches the result of the switch
expression. In \code{async()}, \code{switch} will always either take one of
the given branches or throw an error. Otherwise, the same
conventions (one unnamed argument for a default; fallthrough for
empty arguments) apply as \url{base::switch}.
}
\details{
Coroutine \code{switch} also supports a delimited form of \code{goto}. Within
a branch, \code{goto("other_branch")} will stop executing the present
branch and jump to the new branch named.  Calling \code{goto()} without
arguments will jump back to re-evaluate the switch expression.

If a \code{goto} appears in a \code{try} call, as in

\if{html}{\out{<div class="sourceCode">}}\preformatted{switch("branch",
   branch=try(\{...; goto("otherBranch")\}, finally=\{cleanup()\}),
   otherBranch=... )
}\if{html}{\out{</div>}}

the \code{finally} clause will be executed \emph{before} switching to the new branch.
}
