% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync}
\alias{sync}
\title{Perform a synchronous function call to an asynchronous function}
\usage{
sync(func, ..., .args = list())
}
\arguments{
\item{func}{Asynchronous function to call.}

\item{...}{Arguments to \code{func}.}

\item{.args}{More arguments to \code{func}. This argument is easier to use
programmatically.}
}
\value{
The result of the asynchronous function, i.e. the object that
is passed to its callback. If no results are passed to the callback,
then \code{NULL}.
}
\description{
In case of an asynchronous error, \code{sync} throws the same error.
}
\examples{
result <- sync(
  parallel,
  list(
    asyncify(function() 1L),
    function(callback) callback(NULL, 2L),
    asyncify(function() 3L)
  )
)
result
}
