% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{syncify}
\alias{syncify}
\title{Transform an asynchronous function into a synchronous one}
\usage{
syncify(func)
}
\arguments{
\item{func}{The function to transform.}
}
\value{
The synchronous wrapper function to \code{func}.
}
\description{
All parameters to the wrapper function will be passed to the
asynchronous function, \code{func}.
}
\examples{
sync_parallel <- syncify(parallel)

result <- sync_parallel(list(
  asyncify(function() 1L),
  function(callback) callback(NULL, 2L),
  asyncify(function() 3L)
))
result
}
\seealso{
Other async utilities: \code{\link{async_apply}},
  \code{\link{asyncify}}, \code{\link{reflect}}
}
