% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_for.R
\name{wait_for_all}
\alias{wait_for_all}
\title{Wait for a list of deferred values}
\usage{
wait_for_all(..., .list = list())
}
\arguments{
\item{...}{Deferred values to wait on.}

\item{.list}{More deferred values to wait on.}
}
\description{
Similar to \code{\link[=wait_for]{wait_for()}}, but waits for a list of deferred values.
}
\examples{
urls <- c("https://httpbin.org?q=1", "https://httpbin.org?q=2")
dx <- lapply(urls, http_head)
resp <- wait_for_all(.list = dx)
lapply(resp, "[[", "status_code")
}
\seealso{
Other wait_for functions: \code{\link{wait_for_any}},
  \code{\link{wait_for_env}}, \code{\link{wait_for}}
}
