% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/await.R
\name{wait_for_any}
\alias{wait_for_any}
\title{Wait for any of the deferred values to resolve}
\usage{
wait_for_any(..., .list = list())
}
\arguments{
\item{...}{Deferred values.}

\item{.list}{List of deferred values.}
}
\description{
Returns as soon as any of the specified deferred values is resolved or
rejected. If multiple deferred values resolved, then an arbitrary one
is returned.
}
\examples{
# Returns as soon as one timer expires
t1 <- delay(1)$then(~ 1)
t2 <- delay(1/1000)$then(~ 2)
wait_for_any(t1, t2)
}
\seealso{
Other wait_for functions: \code{\link{wait_for_all}},
  \code{\link{wait_for_env}}, \code{\link{wait_for}}
}
