% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when_all.R
\name{when_all}
\alias{when_all}
\title{Deferred value for a set of deferred values}
\usage{
when_all(..., .list = list())
}
\arguments{
\item{...}{Deferred values.}

\item{.list}{More deferred values.}
}
\value{
A deferred value, that is conditioned on all deferred values
in \code{...} and \code{.list}.
}
\description{
Create a deferred value that is resolved when all listed deferred values
are resolved. Note that the rejection of an input deferred value
triggers the rejection of the deferred value returned by \code{when_all} as
well.
}
\examples{
## Check that the contents of two URLs are the same
u1 <- http_get("https://httpbin.org/get")
u2 <- http_get("https://eu.httpbin.org/get")
dx <- when_all(u1, u2)$
  then(~ identical(.[[1]]$content, .[[2]]$content))
await(dx)
}
\seealso{
\code{\link[=when_any]{when_any()}}
}
