% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when_any.R
\name{when_any}
\alias{when_any}
\title{Resolve a deferred as soon as one deferred in a list resolves}
\usage{
when_any(..., .list = list())
}
\arguments{
\item{...}{Deferred values.}

\item{.list}{More deferred values.}
}
\value{
A deferred value, that is conditioned on all deferred values
in \code{...} and \code{.list}.
}
\description{
Create a deferred value that is resolved as soon as one deferred value
resolves.
}
\examples{
## Use the URL that returns first
u1 <- http_get("https://httpbin.org/get")
u2 <- http_get("https://eu.httpbin.org/get")
dx <- when_any(u1, u2)$then(~ .$url)
await(dx)
}
\seealso{
\code{\link[=when_all]{when_all()}}
}
