% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when_any.R
\name{when_some}
\alias{when_some}
\alias{when_any}
\title{Resolve a deferred as soon as some deferred from a list resolve}
\usage{
when_some(count, ..., .list = list(), cancel = TRUE)

when_any(..., .list = list(), cancel = TRUE)
}
\arguments{
\item{count}{Number of deferred values that need to resolve.}

\item{...}{Deferred values.}

\item{.list}{More deferred values.}

\item{cancel}{Whether to cancel the deferred computations that are
not needed to finish \code{when_some()} or \code{when_any()}, including the
case when one of them throws an error.}
}
\value{
A deferred value, that is conditioned on all deferred values
in \code{...} and \code{.list}.
}
\description{
\code{when_some} creates a deferred value that is resolved as soon as the
specified number of deferred values resolve.
}
\details{
\code{when_any} is a special case for a single.

If the specified number of deferred values cannot be resolved, then
the returned deferred value is rejected.
}
\examples{
## Use the URL that returns first
afun <- function() {
  u1 <- http_get("https://eu.httpbin.org")
  u2 <- http_get("https://eu.httpbin.org/get")
  when_any(u1, u2)$then(~ .$url)
}
synchronise(afun())
}
\seealso{
\code{\link[=when_all]{when_all()}}
}
