% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whilst.R
\name{whilst}
\alias{whilst}
\title{Repeatedly call task, while test returns true}
\usage{
whilst(test_function, task, callback)
}
\arguments{
\item{test_function}{Synchronous truth test to perform before each
execution of task. Invoked without arguments.}

\item{task}{An async function which is called each time test
passes. Invoked with \code{(callback)}.}

\item{callback}{A callback which is called after the test function has
failed and repeated execution of \code{task} has stopped. \code{callback} will
be passed an error and any arguments passed to the final task's
callback. Invoked with \code{(err, [results])}.}
}
\value{
Task id.
}
\description{
Calls callback when stopped, or an error occurs.
}
\examples{
## Keep calling while result is bigger than 0.1
calls <- 0
number <- Inf
wait_for(whilst(
  function() number > 0.1,
  function(callback) {
    number <<- runif(1)
    calls <<- calls + 1
    callback(NULL, number)
  },
  function(err, res) print(res)
))
calls
}
\seealso{
Other async control flow: \code{\link{parallel_limit}},
  \code{\link{parallel}}, \code{\link{retryable}},
  \code{\link{retry}}, \code{\link{sequence}},
  \code{\link{until}}, \code{\link{waterfall}}
}
