% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async-description.R
\name{async-package}
\alias{async-package}
\title{Asynchronous code constructs.}
\description{
\itemize{
\item{\code{\link{gen}} and \code{\link{yield}} create generators (iterators) based on sequential code}
\item{\code{\link{async}} and \code{\link{await}} create asynchronously computed values (promises) based on sequential code.}
}
}
\author{
Peter Meilstrup
}
