% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R, R/package.R
\docType{package}
\name{async}
\alias{async}
\alias{async}
\alias{async-package}
\title{Create an async function}
\usage{
async(fun)
}
\arguments{
\item{fun}{Original function.}
}
\value{
Async version of the original function.
}
\description{
Create an async function, that returns a deferred value, from a
regular function. If \code{fun} is already an async function, then it does
nothing, just returns it.

Asynchronous HTTP
}
\details{
The result function will have the same arguments, with the same default
values, and the same environment as the original input function.
}
\section{Introduction}{


The \code{async} package brings asynchronous computation and I/O to R.

It has an event loop that takes care of scheduling multiple asynchronous
functions, and polling for input from HTTP and other sources.

TODO: more
}

\examples{
f <- function(x) 42
af <- async(f)
is_async(f)
is_async(af)
f()
dx <- af()
dx
await(dx)
}
