% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{async}
\alias{async}
\alias{async-package}
\title{Asynchronous HTTP}
\description{
Asynchronous HTTP
}
\section{Introduction}{


The \code{async} package brings asynchronous computation and I/O to R.

It has an event loop that takes care of scheduling multiple asynchronous
functions, and polling for input from HTTP and other sources.

TODO: more
}

