% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{async}
\alias{async}
\alias{async-package}
\title{Asynchronous HTTP}
\description{
The \code{async} package brings asynchronous computation to R.
Currently it supports asynchronous HTTP operations, and timers.
(Other primitive asynch operations will be added in the future.)
}
\section{Async Functions}{


Asynchronous computation happens in special functions called tasks.
A task is a function that reports its completion asynchronously, via
calling a callback function.

TODO: more
}

\section{List iterators}{


The package also provides several utilities to make it easier to
program with tasks and avoid deeply nested callbacks.

TODO: more
}

\section{Synchronization}{


TODO
}

