% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{async}
\alias{async}
\title{R6 Class async}
\description{
simple tool to pass message of progress and interruption of routines
between some R processes like shiny reactives and futures routines.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{status_file}}{path to temporary file}

\item{\code{upper}}{The value that represents the end of the routine progress.}

\item{\code{lower}}{The value that represents the starting
point of the routine progress.}

\item{\code{auto.finish}}{if it is true the routine will
be interrupted when the progress is equal to the upper value.}

\item{\code{value_message}}{get job status.}

\item{\code{reactive}}{If its true generate a reactive expression
is a expression whose result will change over time.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-async-new}{\code{async$new()}}
\item \href{#method-async-interrupt}{\code{async$interrupt()}}
\item \href{#method-async-set}{\code{async$set()}}
\item \href{#method-async-inc}{\code{async$inc()}}
\item \href{#method-async-status}{\code{async$status()}}
\item \href{#method-async-check}{\code{async$check()}}
\item \href{#method-async-finalize}{\code{async$finalize()}}
\item \href{#method-async-clone}{\code{async$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-async-new"></a>}}
\if{latex}{\out{\hypertarget{method-async-new}{}}}
\subsection{Method \code{new()}}{
create an interactive environment
to pass progress message between
R processes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$new(
  lower,
  upper,
  auto.finish = TRUE,
  reactive = TRUE,
  verbose = FALSE,
  msg,
  detail
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lower}}{The value that represents the starting
point of the routine progress.}

\item{\code{upper}}{upper The value that represents
the end of the routine progress.}

\item{\code{auto.finish}}{auto.finish if it is true the routine will
be interrupted when the progress is equal to the upper value.}

\item{\code{reactive}}{If its true generate a reactive expression
is a expression whose result will change over time.}

\item{\code{verbose}}{logical.If its TRUE  provides much more
information about the flow of information between the R processes.}

\item{\code{msg}}{A single-element character vector;
the initial message to be pass between processes.}

\item{\code{detail}}{A single-element character vector;
the initial detail message to be pass between processes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-async-interrupt"></a>}}
\if{latex}{\out{\hypertarget{method-async-interrupt}{}}}
\subsection{Method \code{interrupt()}}{
Interrupt the process tracked by the async function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$interrupt(msg = "process interrupted", detail = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{A single-element character vector;
the interrupt message to be pass between processes.}

\item{\code{detail}}{A single-element character vector;
the interrupt detail message to be pass between processes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-async-set"></a>}}
\if{latex}{\out{\hypertarget{method-async-set}{}}}
\subsection{Method \code{set()}}{
Set progress of the tracked routine.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$set(value = 0, msg = "Running...", detail = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{the value at which to set the progress,
relative to lower and upper.}

\item{\code{msg}}{A single-element character vector;
the message to be pass between processes.}

\item{\code{detail}}{A single-element character vector;
the detail message to be pass between processes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-async-inc"></a>}}
\if{latex}{\out{\hypertarget{method-async-inc}{}}}
\subsection{Method \code{inc()}}{
Increment the progress of the tracked routine.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$inc(value = 0, msg = "Running...", detail = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{the value at which to set the progress,
relative to lower and upper.}

\item{\code{msg}}{A single-element character vector;
the message to be pass between processes.}

\item{\code{detail}}{A single-element character vector;
the detail message to be pass between processes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-async-status"></a>}}
\if{latex}{\out{\hypertarget{method-async-status}{}}}
\subsection{Method \code{status()}}{
get the status of progress
out of the context being tracked.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-async-check"></a>}}
\if{latex}{\out{\hypertarget{method-async-check}{}}}
\subsection{Method \code{check()}}{
Check the status process in
the tracked context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$check()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-async-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-async-finalize}{}}}
\subsection{Method \code{finalize()}}{
Close all routines of async tracking.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-async-clone"></a>}}
\if{latex}{\out{\hypertarget{method-async-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
