% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{async}
\alias{async}
\title{R6 Class async}
\description{
simple tool to pass message of progress and interruption of routines
between some R processes like shiny reactives and futures routines.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{If}}{its true generate a reactive expression
is a expression whose result will change over time.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{status_file}}{path to temporary file}

\item{\code{upper}}{The value that represents the end of the routine progress.}

\item{\code{lower}}{The value that represents the starting
point of the routine progress.}

\item{\code{auto.finish}}{if it is true the routine will
be interrupted when the progress is equal to the upper value.}

\item{\code{value_message}}{get job status.}

\item{\code{If}}{its true generate a reactive expression
is a expression whose result will change over time.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{async$new()}}
\item \href{#method-interrupt}{\code{async$interrupt()}}
\item \href{#method-progress}{\code{async$progress()}}
\item \href{#method-status}{\code{async$status()}}
\item \href{#method-check}{\code{async$check()}}
\item \href{#method-clone}{\code{async$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
create an interative environment
to pass progress message between
R processes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$new(lower = 0, upper = 100, auto.finish = TRUE, reactive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lower}}{The value that represents the starting
point of the routine progress.}

\item{\code{upper}}{upper The value that represents
the end of the routine progress.}

\item{\code{auto.finish}}{auto.finish if it is true the routine will
be interrupted when the progress is equal to the upper value.}

\item{\code{reactive}}{If its true generate a reactive expression
is a expression whose result will change over time.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-interrupt"></a>}}
\if{latex}{\out{\hypertarget{method-interrupt}{}}}
\subsection{Method \code{interrupt()}}{
Interrupt checking to future processes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$interrupt(msg = "process interrupted")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msg}}{A single-element character vector;
the message to be displayed to the user.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-progress"></a>}}
\if{latex}{\out{\hypertarget{method-progress}{}}}
\subsection{Method \code{progress()}}{
Set progress to future processes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$progress(value = 0, msg = "Running...")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{the value at which to set the progress,
relative to lower and upper.}

\item{\code{msg}}{A single-element character vector;
the message to be displayed to the user.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-status"></a>}}
\if{latex}{\out{\hypertarget{method-status}{}}}
\subsection{Method \code{status()}}{
get the status of the process out
of the future context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check"></a>}}
\if{latex}{\out{\hypertarget{method-check}{}}}
\subsection{Method \code{check()}}{
Check the status of the process in
the future context
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$check()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
