% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async_bar.R
\name{async_bar}
\alias{async_bar}
\title{R6 Class async_bar}
\description{
Reports  async class progress with a shiny bar to end user.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{async_bar$new()}}
\item \href{#method-progress}{\code{async_bar$progress()}}
\item \href{#method-finalize}{\code{async_bar$finalize()}}
\item \href{#method-clone}{\code{async_bar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Interactive environment
to create a progress  bar of async tracking.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async_bar$new(async, id, interval = 400, max.rep = 20, detail = "", session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{async}}{R6 class. Object to tracking routines.}

\item{\code{id}}{character. ID of progress bar.}

\item{\code{interval}}{numeric. Approximate number of milliseconds
to wait between checks of the job progress.}

\item{\code{detail}}{detail message to display.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-progress"></a>}}
\if{latex}{\out{\hypertarget{method-progress}{}}}
\subsection{Method \code{progress()}}{
observe event to checks of the async job progress.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async_bar$progress(session, input)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{shiny session}

\item{\code{input}}{shiny input}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
Close all routines of async bar.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async_bar$finalize(session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{shiny session}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{async_bar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
