% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{async_filter}
\alias{async_filter}
\title{Keep element using an asyncronous predicate function}
\usage{
async_filter(.x, .p, ...)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.p}{An asynchronous predicate function.}

\item{...}{Additional arguments to the predicate function.}
}
\value{
A deferred value for the result.
}
\description{
Keep element using an asyncronous predicate function
}
\examples{
## Filter out non-working URLs
urls <- c("https://httpbin.org/get", "https://httpbin.org/status/404")
test_url <- async_sequence(http_head, ~ identical(.$status_code, 200L))
await(async_filter(urls, test_url))
}
\seealso{
Other async iterators: \code{\link{async_detect}},
  \code{\link{async_every}}, \code{\link{async_map}}
}
\concept{async iterators}
