% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/await.R
\name{await}
\alias{await}
\title{Wait for the resolution of deferred a value}
\usage{
await(def)
}
\arguments{
\item{def}{The deferred value to wait on. If not a deferred value,
then it is resolved immediately, to itself.}
}
\value{
The value of the deferred.
}
\description{
Either until it is successfully resolved, or it is rejected.
While waiting for the deferred value to resolve, the async event loop
is running, so other async operations do \emph{not} stop.
}
\details{
If the deferred value resolves to another deferred value, then
it wait on that as well, and so forth.
}
\examples{
dx <- delay(1/10)$then(~ 42)
dx
await(dx)
await("foobar")
}
\seealso{
Other await functions: \code{\link{await_all}},
  \code{\link{await_any}}, \code{\link{await_env}}
}
\concept{await functions}
