% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/await.R
\name{await_any}
\alias{await_any}
\title{Wait for any of the deferred values to resolve}
\usage{
await_any(..., .list = list())
}
\arguments{
\item{...}{Deferred values.}

\item{.list}{List of deferred values.}
}
\description{
Returns as soon as any of the specified deferred values is resolved or
rejected. If multiple deferred values resolved, then an arbitrary one
is returned.
}
\examples{
# Returns as soon as one timer expires
t1 <- delay(1)$then(~ 1)
t2 <- delay(1/1000)$then(~ 2)
await_any(t1, t2)
}
\seealso{
Other await functions: \code{\link{await_all}},
  \code{\link{await_env}}, \code{\link{await}}
}
\concept{await functions}
