% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect.channel}
\alias{collect.channel}
\title{Collect all values emitted by a channel.}
\usage{
collect.channel(ch, type = list())
}
\arguments{
\item{ch}{a \link{channel} object.}

\item{type}{Optionally provide a vector of the desired output type
(similarly to using \code{vapply}); defaults to \code{list()}}
}
\description{
\code{collect.channel} takes a \link{channel} as argument and returns a
\link{promise}.  It behaves as an analogue of \link{as.list.iteror}; all
values emitted by the channel will be collected into a list (or
other vector.) When the source channel closes, the promise will
resolve with the collected list.
}
