% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coroutine.R
\name{debugAsync}
\alias{debugAsync}
\title{Toggle single-step debugging for a coroutine.}
\usage{
debugAsync(x, R, internal, trace)
}
\arguments{
\item{x}{A coroutine object as constructed by (\link{async}, \link{gen} or
\link{stream}).}

\item{R}{Set TRUE to step through expressions at user level}

\item{internal}{Set TRUE to step through at coroutine
implementation level.}

\item{trace}{Set TRUE or provide a print function to print each R
expression evaluated in turn.}
}
\value{
a \code{list(R=, internal=, trace=)} with the current debug state.
}
\description{
Toggle single-step debugging for a coroutine.
}
