% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cps.r
\name{delay}
\alias{delay}
\title{"delay" returns a promise which resolves only after the specified
number of seconds. This uses the R event loop via [later].}
\usage{
delay(delay, expr = NULL)
}
\arguments{
\item{delay}{The promise will resolve after at least this many seconds.}

\item{expr}{The value to resolve with; will be forced after the delay.}
}
\description{
"delay" returns a promise which resolves only after the specified
number of seconds. This uses the R event loop via [later].
}
\examples{
# print a message after a few seconds
async({await(delay(10)); cat("Time's up!\n")})
}
