% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay.R
\name{delay}
\alias{delay}
\title{Run a function after the specified time interval}
\usage{
delay(delay)
}
\arguments{
\item{delay}{Time interval in seconds, the amount of time to delay
to delay the execution of the callback. It can be a fraction of a
second.}
}
\value{
A deferred object.
}
\description{
Since R is single-threaded, the callback might be executed (much) later
than the specified time period.
}
\examples{
## Two HEAD requests with 1/2 sec delay between them
resp <- list()
dx <- http_head("https://httpbin.org?q=2")$
  then(function(value) resp[[1]] <<- value$status_code)$
  then(function(...) delay(1/2))$
  then(function(...) http_head("https://httpbin.org?q=2"))$
  then(function(value) resp[[2]] <<- value$status_code)
await(dx)
resp
}
