% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{http_get}
\alias{http_get}
\title{Asynchronous HTTP GET request}
\usage{
http_get(url)
}
\arguments{
\item{url}{URL to connect to.}
}
\value{
Deferred object.
}
\description{
Start an HTTP GET request in the background, and report its completion
via a deferred.
}
\examples{
dx <- http_get("https://httpbin.org/status/200")$
  then(~ .$status_code)
await(dx)
}
\seealso{
Other asyncronous HTTP calls: \code{\link{http_head}}
}
