% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{http_head}
\alias{http_head}
\title{Asynchronous HTTP HEAD request}
\usage{
http_head(url)
}
\arguments{
\item{url}{URL to connect to.}
}
\value{
Deferred object.
}
\description{
Asynchronous HTTP HEAD request
}
\examples{
dx <- http_head("https://httpbin.org/status/200")$
  then(~ .$status_code)
await(dx)

# Check a list of URLs in parallel
urls <- c("https://r-project.org", "https://httpbin.org")
dx <- when_all(.list = lapply(urls, http_head))$
  then(~ lapply(., "[[", "status_code"))
await(dx)
}
\seealso{
Other asyncronous HTTP calls: \code{\link{http_get}}
}
