% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{http_head}
\alias{http_head}
\title{Asynchronous HTTP HEAD request}
\usage{
http_head(url, headers = character(), file = NULL, timeout = 10,
  on_progress = NULL)
}
\arguments{
\item{url}{URL to connect to.}

\item{headers}{HTTP headers to send.}

\item{file}{If not \code{NULL}, it must be a string, specifying a file.
The body of the response is written to this file.}

\item{timeout}{Timeout for the request, in seconds.}

\item{on_progress}{Progress handler function. It is only used if the
response body is written to a file.}
}
\value{
Deferred object.
}
\description{
Asynchronous HTTP HEAD request
}
\examples{
afun <- async(function() {
  dx <- http_head("https://eu.httpbin.org/status/200")$
    then(~ .$status_code)
})
sync_wrap(afun())

# Check a list of URLs in parallel
afun <- async(function(urls) {
  when_all(.list = lapply(urls, http_head))$
    then(~ lapply(., "[[", "status_code"))
})
urls <- c("https://r-project.org", "https://eu.httpbin.org")
sync_wrap(afun(urls))
}
\seealso{
Other asyncronous HTTP calls: \code{\link{http_get}}
}
\concept{asyncronous HTTP calls}
