% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{http_head}
\alias{http_head}
\title{Asynchronous HTTP HEAD request}
\usage{
http_head(url, headers = character(), file = NULL, on_progress = NULL)
}
\arguments{
\item{url}{URL to connect to.}

\item{headers}{HTTP headers to send.}

\item{file}{If not \code{NULL}, it must be a string, specifying a file.
The body of the response is written to this file.}

\item{on_progress}{Progress handler function. It is only used if the
response body is written to a file.}
}
\value{
Deferred object.
}
\description{
Asynchronous HTTP HEAD request
}
\examples{
dx <- http_head("https://httpbin.org/status/200")$
  then(~ .$status_code)
wait_for(dx)

# Check a list of URLs in parallel
urls <- c("https://r-project.org", "https://httpbin.org")
dx <- when_all(.list = lapply(urls, http_head))$
  then(~ lapply(., "[[", "status_code"))
wait_for(dx)
}
\seealso{
Other asyncronous HTTP calls: \code{\link{http_get}}
}
