% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{http_post}
\alias{http_post}
\title{Asynchronous HTTP POST request}
\usage{
http_post(
  url,
  data = NULL,
  headers = character(),
  file = NULL,
  options = list(),
  form_data = NULL,
  on_progress = NULL
)
}
\arguments{
\item{url}{URL to connect to.}

\item{data}{Data to send. Either a raw vector, or a character string
that will be converted to raw with \link[base:rawConversion]{base::charToRaw}.
Exactly one of \code{data} and \code{form_data} must be given.}

\item{headers}{HTTP headers to send.}

\item{file}{If not \code{NULL}, it must be a string, specifying a file.
The body of the response is written to this file.}

\item{options}{Options to set on the handle. Passed to
\code{\link[curl:handle]{curl::handle_setopt()}}.}

\item{form_data}{Form data to send. A list where each element is named
with the key. Exactly one of \code{data} and \code{form_data} must be given.}

\item{on_progress}{Progress handler function. It is only used if the
response body is written to a file. See details at \code{\link[=http_get]{http_get()}}.}
}
\description{
Start an HTTP POST request in the background, and report its completion
via a deferred value.
}
\examples{
json <- jsonlite::toJSON(list(baz = 100, foo = "bar"))

do <- function() {
  headers <- c("content-type" = "application/json")
  http_post("https://eu.httpbin.org/post", data = json, headers = headers)$
    then(http_stop_for_status)$
    then(function(x) {
      jsonlite::fromJSON(rawToChar(x$content))$json
    })
}

synchronise(do())
}
