% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{http_stop_for_status}
\alias{http_stop_for_status}
\title{Throw R errors for HTTP errors}
\usage{
http_stop_for_status(resp)
}
\arguments{
\item{resp}{HTTP response from \code{\link[=http_get]{http_get()}}, \code{\link[=http_head]{http_head()}}, etc.}
}
\value{
The HTTP response, invisibly, if it is considered successful.
Otherwise an error is thrown.
}
\description{
Status codes below 300 are considered successful, others will trigger
errors.
}
\examples{
dx <- http_get("https://httpbin.org/status/404")$
  then(http_stop_for_status)

tryCatch(await(dx), error = function(e) e)
}
