% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteror.R
\name{iteror}
\alias{iteror}
\title{An efficient and succinct iteration protocol}
\usage{
iteror(it, ...)
}
\description{
The "iterators" package uses stop("StopIteration") and tryCatch to
signal end of iteration, but tryCatch has a not-insiginificant
amount of overhead. In the context of a generator, when you are in a
}
